[ELG-189] Fix. Hold AP rssi in int16_t not int8_t
Update sample_client to reflect this fix also

+++ libel/beacons.h
@@ -46,7 +46,7 @@ struct ap {
     uint8_t mac[MAC_SIZE];
     uint32_t age;
     uint32_t freq;
-    int8_t rssi;
+    int16_t rssi;
 };
 
 // http://wiki.opencellid.org/wiki/API
